requirements.txt
numpy>=1.24.0,<2.0.0
pandas>=2.0.0,<3.0.0
Pillow>=8.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
google-cloud-storage>=2.0.0
torch>=1.12.0
torchvision>=0.13.0
scikit-learn>=1.0.0
imbalanced-learn>=0.10.0
joblib>=1.1.0
pathlib2>=2.3.0 